Add files via upload
The name of the game is “Unseal Password”. In this game you will be having a virtual experience of the present situation which the world is facing. Yeah, you have guessed it right, it’s COVID-19. In this game you have to save the girl from the viruses and there are several boosters like masks and sanitizers. 
It is a single level game but there is game adaptation so, it’s quite challenging for the players to score high. As the score increases the path will move faster and all the object’s speed will also increase.
RULES
1. Mask=10points
2. 1 Sanitizer=50points
3. If the girl touches the virus, then it’s game over.